#8:
- created Pages folder
- outsource Sign Up code for reusable components
- created Login Page
