+++ src/contexts/AuthContext.tsx
@@ -14,6 +14,7 @@ import { User } from "firebase/auth";
  */
 interface AuthContextProps {
   user: User | null;
+  login: (email: string, password: string) => Promise<void>;
   signup: (
     email: string,
     password: string,
